run linting
