+++ test/unit/component/parser/json/BpmnJsonParser.callActivity.test.ts
@@ -22,7 +22,7 @@ import { verifyEdge, verifyShape } from '../../../helpers/bpmn-model-expect';
 
 import type { GlobalTaskKind } from '@lib/model/bpmn/internal';
 import { ShapeBpmnCallActivityKind, ShapeBpmnElementKind, ShapeBpmnMarkerKind } from '@lib/model/bpmn/internal';
-import type { BpmnJsonModel } from '@lib/model/bpmn/json/BPMN20';
+import type { BpmnJsonModel } from '@lib/model/bpmn/json/bpmn20';
 
 const expectedBounds = { x: 346, y: 856, width: 45, height: 56 };
