+++ test/unit/component/parser/json/BpmnJsonParser.diagram.test.ts
@@ -18,7 +18,7 @@ import { parseJsonAndExpectOnlyFlowNodes } from '../../../helpers/JsonTestUtils'
 import { verifyShape } from '../../../helpers/bpmn-model-expect';
 
 import { ShapeBpmnElementKind } from '@lib/model/bpmn/internal';
-import type { BpmnJsonModel } from '@lib/model/bpmn/json/BPMN20';
+import type { BpmnJsonModel } from '@lib/model/bpmn/json/bpmn20';
 
 describe('parse bpmn as json for diagram', () => {
   const jsonModelWithoutBPMNDiagram: BpmnJsonModel = {
