--- test/unit/component/parser/xml/BpmnXmlParser.bonita-community-2021_1.test.ts
@@ -1,42 +1,0 @@
-/*
-Copyright 2021 Bonitasoft S.A.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-import BpmnXmlParser from '@lib/component/parser/xml/BpmnXmlParser';
-import type { TProcess } from '@lib/model/bpmn/json/baseElement/rootElement/rootElement';
-import type { BPMNDiagram } from '@lib/model/bpmn/json/BPMNDI';
-import { readFileSync } from '@test/shared/file-helper';
-
-describe('parse bpmn as xml for Bonita Community 2021.1', () => {
-  it('bpmn with process with extension, ensure elements are present', () => {
-    const a20Process = readFileSync('../fixtures/bpmn/xml-parsing/bonita-community-2021.1-A.2.0.export.bpmn');
-
-    const json = new BpmnXmlParser().parse(a20Process);
-
-    // Bonita generates a single object, most Modeler generates an array with a single element
-    const process: TProcess = json.definitions.process as TProcess;
-    expect(process.id).toBe('_GW8mAGJcEeuag9wrqgia1Q');
-    expect(process.name).toBe('WFP-6-');
-
-    expect(process.task).toHaveLength(4);
-    expect(process.exclusiveGateway).toHaveLength(2);
-    expect(process.sequenceFlow).toHaveLength(9);
-
-    const bpmnDiagram: BPMNDiagram = json.definitions.BPMNDiagram as BPMNDiagram;
-    expect(bpmnDiagram.BPMNPlane.BPMNShape).toHaveLength(9);
-    expect(bpmnDiagram.BPMNPlane.BPMNEdge).toHaveLength(9);
-    expect(bpmnDiagram.BPMNLabelStyle).toBeDefined();
-  });
-});
